% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacckFillLeadingNas.R
\name{bacckFillLeadingNas}
\alias{bacckFillLeadingNas}
\title{Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month}
\usage{
bacckFillLeadingNas(datain, f = 12)
}
\arguments{
\item{datain}{a time series matrix with lags as each column, contained leading NAs}

\item{f}{frequency of the time series}
}
\value{
a data frame with the leading NAs filled in
}
\description{
Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month
}
\details{
this is too clean up lagged regressors for a xreg argument of an arima model
}
\author{
Zeng, Qichen
}

