% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmapper.R
\name{calendar.period.mapper}
\alias{calendar.period.mapper}
\title{This function creates a vector of length n of calendar periods}
\usage{
calendar.period.mapper(st, n, useFirst = TRUE)
}
\arguments{
\item{st}{The start of the calendar periods, an integer fiscal period of the form yyyymm}

\item{n}{The integer number of fiscal periods in the map}

\item{useFirst}{Logical where or not to include the 'st' variable a in the output vector}
}
\value{
A vector of integet fiscal periods
}
\description{
This function creates a vector of length n of calendar periods
}
\details{
The output is of length n, reguardless of where the 'st' period is included
}
\author{
Matthew Davis
}

