% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPtoCalendar.R
\name{CPtoCalendar}
\alias{CPtoCalendar}
\title{Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month}
\usage{
CPtoCalendar(mydata, fillgap = TRUE, fiscal = TRUE,
  calendarname = holidayNYSE(2000:2020))
}
\arguments{
\item{mydata}{an integer vector of calendar periods of the form yyyymm (defaults to fiscal)}

\item{fillgap}{whether to fill in any missing fiscal periods, or only use the list given}

\item{fiscal}{whether the calendar periods are a fiscal or calendar period}

\item{calendarname}{Name for timdate package to deterimine holidays na buisness days}
}
\value{
a data frame of sums of weekdays, bizdays and holidays by the fiscal periods entered
}
\description{
Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month
}
\details{
This package is intended to get montly summaries of number holidays and weekdays for a particular calenar or fiscal month.
}
\author{
Qichen Zeng, Matthew Davis
}

