% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp2cp.R
\name{FPtoCP}
\alias{FPtoCP}
\title{Converts a Fiscal Period to to a Caendar Period to a Fiscal Period}
\usage{
FPtoCP(fp, n = 6, reverse = FALSE)
}
\arguments{
\item{fp}{an integer vector of fiscal period of the form yyyymm}

\item{n}{number of months difference between the fiscal year and the calendar year}

\item{reverse}{True or FALSE reverses the function to go from CP to FP}
}
\value{
A vector of calendar periods  of the form yyyymm
}
\description{
Converts a Fiscal Period to to a Caendar Period to a Fiscal Period
}
\details{
A fiscal period of 201701 starting in July CY 2016 would have a n=6 (
This is 'new fiscal year starts july 1st' scenario
}
\author{
Matthew Davis
}

