% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp2date.R
\name{fptoDate}
\alias{fptoDate}
\title{Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month}
\usage{
fptoDate(mydata, n = 6, fiscal = TRUE)
}
\arguments{
\item{mydata}{an integer vector of fiscal period of the form yyyymm}

\item{n}{number of months difference between the fiscal year and the calendar year}

\item{fiscal}{Logical, whether the mydata consisits of calendar or fiscal periods}
}
\value{
a vector of dates, where the day is the first day of the month
}
\description{
Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month
}
\details{
A fiscal period 201701 starting in July 1st, 2016 would have a n=6 (
This is 'new fiscal year starts july 1st' scenario.  201701 should return as.Date('2016-01-07')
}
\author{
Matthew Davis
}

