% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gowthMetric.R
\name{growthMetric}
\alias{growthMetric}
\title{Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month}
\usage{
growthMetric(myTs, fiscal = TRUE)
}
\arguments{
\item{myTs}{a univariate time series object of frequeny = 12 (monthly)}

\item{fiscal}{Logical, whether to return fiscal or calendar metrics (FY start July 1st)}
}
\value{
a dataframe of growth metrics
}
\description{
Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month
}
\details{
Year over Year 'yoy' is the percent change over the pervious year,
}
\author{
Matthew Davis
}

