% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longFormPreformance.R
\name{longFormPreformance}
\alias{longFormPreformance}
\title{Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month}
\usage{
longFormPreformance(mydata, nameKey = "FinancialSubdivisionName",
  valueCol = "PaymentAmount", timeCol = "CalendarPeriod",
  forcastInicatorCol = "forecastIncator")
}
\arguments{
\item{mydata}{data frame with actual and forecasted data in long form}

\item{nameKey}{column name that is the primary key for the forecast (ie bottom level)}

\item{valueCol}{column names that has the forecats and actuals vales}

\item{timeCol}{column with the time reference (intersection become the validation set)}

\item{forcastInicatorCol}{logical, TRUE ==  is a forecasted value, FALSE = is actyual}
}
\value{
a dataframe of preformance metrics
}
\description{
Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month
}
\author{
Matthew Davis
}

