% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanAdjustForecast.R
\name{meanAdjustForecast}
\alias{meanAdjustForecast}
\title{Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month}
\usage{
meanAdjustForecast(rd.ts, fd.ts)
}
\arguments{
\item{rd.ts}{Time Series Matrix with raw data}

\item{fd.ts}{Time Series Matrix of}
}
\value{
a dataframe of growth metrics
}
\description{
Takes a vector of fiscal or calendar periods and returns a vector of dates, the first date of each month
}
\details{
Year over Year 'yoy' is the percent change over the pervious year,
}
\author{
Matthew Davis
}

