% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPreformance.R
\name{modelPreformance}
\alias{modelPreformance}
\title{calculated a bunch of model preformance metrics.}
\usage{
modelPreformance(a, p)
}
\arguments{
\item{a}{vector of actuals}

\item{p}{vector of predictions, same length as actuals}
}
\value{
a dataframe of preformance measurs
}
\description{
calculated a bunch of model preformance metrics.
}
\details{
The following are the returned metrics
\itemize{
\item cor = pairwise complete,
\item  r2 = cor^2 ie percent of variance explained
\item rmse = root mean squared error
\item mse = mean squared error
\item pvalue =  from paired T Test
\item  meanA =  Actuals na.rm
\item meanP = Prediction na.rm
\item normalResids = p value from shapiro test, to test whether resids are normally distrubuted
}
}
\author{
Matthew Davis
}

