% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFPreformance.R
\name{plotForecastPreformance}
\alias{plotForecastPreformance}
\title{Calculate Forecast Model Preformance}
\usage{
plotForecastPreformance(fcastData, trainData, seriesName = "")
}
\arguments{
\item{fcastData}{univariate ts with the forecasted data (must overlap trainData for preformance metric calcs)}

\item{trainData}{univariate ts with the training data}

\item{seriesName}{name ofthe time series}
}
\value{
a plot of preformance from the time series
}
\description{
plots a forecast and training data from time series objects
}
\details{
This is intended to plot and look and anualized time series forecast and  estimate preformance
}
\author{
Matthew Davis
}

