% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsDensities.R
\name{plotTsDenisities}
\alias{plotTsDenisities}
\title{plot time series density before and after}
\usage{
plotTsDenisities(myTs, timeCut = c(2016), plotName = NULL)
}
\arguments{
\item{myTs}{a univariate time series object}

\item{timeCut}{numeric date date to cut test before and after ie 2016.5}

\item{plotName}{name of the plot to be put in the title}
}
\value{
density plot of the data before and after the cut off
}
\description{
this function is used to plot densities before and after a certain date
}
\details{
This creates a ts matrix of lag columns from a ts object
}
\author{
Matthew Davis
}

