% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2lagMatrix.R
\name{ts2lagMatrix}
\alias{ts2lagMatrix}
\title{Build Time Series Lag Matrix}
\usage{
ts2lagMatrix(myTs, nlag = 1, pastLags = TRUE)
}
\arguments{
\item{myTs}{a univariate time series object}

\item{nlag}{number of lags}
}
\value{
time series object where each column is a lag of the original data
}
\details{
This creates a ts matrix of lag columns from a ts object
}
\author{
Matthew Davis
}

